rename KeybindsManager to PlayerSettingsManager since unities built in scripts handle that. new script will contain settings for player like decidign if abilities will be auto casted where possible. or fully manual. (atm just that for now).
added bool manualCastOffensiveAbilities to PlayerData in SaveManager and to PlayerSettingsManager. linked both together. made new ui panel to contail these new settings. + a way to toggle it and see what its currently set to.

updated PlayerController script to check these bools in static instance of PlayerSettingsManager to do differnt things based on what the bool is set to.

spent a few hrs debugging odd behaviour with context menu and unequipping items. (stats not updating, abilities being sent to inventory, null ref exc on unequipping items from hotbar for a second time but item in slot not being null in inspector but debugging as null, and the intended action going through for some items but not for others)

tracking down issue with onClick listners and forgetting to add remove all listners from unEquipItemButton took th emajority of the time. also fixed stats not updating in ui with oldSlot.RemoveItemFromSlot().
abilities are still sent to inventory and not properly unequipped, that ill get fixed.