added call to unload networked scene for joining clients when they disconnect.
added helper functions to EntityClassHandler called GetIndexOfClass and GetIndexOfStatBoost.

added code to now sync player info of clients already in the game to newly joined clients.

once a new client connects and there local player ref is set, an rpc call to server/host is made for every other player obj that isnt them. the host will then send an rpc back to the client that sent it with the rest of the players info.

TODO:
add in stat boosts to be fetched by the same rpc, test again to make sure it works.
move onto syncing player equipping/unequipping of items
then make them sync with already connected clients