spent another 2 hrs running different tests and finally tracked down the cause of projectiles being duped. the temp unique ranged weapon SO that slimes use kept duping weapon projectiles leading to desyncing of visual attacks of slimes + projectiles slowly breaking further.
issue is due to the SO Unique ranged weapon not being in the asset database so when it comes to syncing the weaponBaseRef variable in Projectiles script it fails. debug log confirms it.

added error logs for the future notify of the issue as projectiles need that ref, updating and restoring any changes made whilst testing. small rewrite of how entity abilities are randomly chosen. + added any missing weapons etc to database for now.